区分文件夹/文件配置
