+++ package.nls.zh-cn.json
@@ -15,9 +15,10 @@
   "text.config.defaultFolderTemplate": "创建文件夹时默认使用的模板",
   "text.config.defaultFileTemplate": "创建文件时默认使用的模板",
   "text.templateListItemDetail": "模板来源: ${1}配置",
+  "text.templateListItemDetailWithType": "${1}模板: ${2}配置",
   "text.templateListItemPlaceholder": "选择一个模板开始创建",
-  "text.templateItemType.folder": "文件夹",
-  "text.templateItemType.file": "文件",
+  "text.templateItemType.folders": "文件夹",
+  "text.templateItemType.files": "文件",
 
   "text.success.create": "Create: 创建 ${1} 成功",
