+++ src/commands/create-file.ts
@@ -1,7 +1,7 @@
 import vscode from 'vscode'
 
 import { WORKSPACE_PATH, localize, log } from '../utils'
-import { getMenuRelativePath } from '../core'
+import { getMenuRelativePath, expandTemplateItems, showTemplateList } from '../core'
 
 export function createFile() {
   console.log('createFile')
@@ -15,6 +15,17 @@ export function registerCreateFile() {
     }
 
     const initPath = getMenuRelativePath(uri)
+
+    const templateItems = expandTemplateItems('files', { initPath })
+
+    showTemplateList(templateItems).then(res => {
+      console.log(res)
+      // if (typeof res.renderer !== 'function') {
+      //   const errorMessage = localize.getLocalize('text.error.templateFunction', res.label)
+      //   log.error(errorMessage, true)
+      //   return Promise.reject(errorMessage)
+      // }
+    })
 
     console.log(initPath)
   })
